Fixed warnings when compiling for x64 on MSVC through VS 2017 v15.3.3.

+++ code/glTF2Asset.inl
@@ -242,7 +242,7 @@ Ref<T> LazyDict<T>::Create(const char* id)
     }
     T* inst = new T();
     inst->id = id;
-    inst->index = mObjs.size();
+    inst->index = static_cast<int>(mObjs.size());
     return Add(inst);
 }
